requirements.txt
python-telegram-bot==13.15
urllib3==1.26.16
standard-imghdr
